feat. User테이블 추가 및 User, Auth 폴더 생성

+++ src/surveys/services/surveys.service.ts
@@ -4,6 +4,7 @@ import {
   Injectable,
   Logger,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateSurveyDto } from '../dto/create-survey.dto';
@@ -13,6 +14,7 @@ import { Repository } from 'typeorm';
 import { Surveys } from '../entities/surveys.entity';
 import { EntityWithId } from 'src/survey.type';
 import { Questions } from 'src/questions/entities/questions.entity';
+import { Users } from 'src/users/entities/user.entity';
 
 @Injectable()
 export class SurveysService {
@@ -79,7 +81,10 @@ export class SurveysService {
   }
 
   // 설문지 생성 (createSurvey)
-  async createSurvey(createDto: CreateSurveyDto): Promise<Surveys> {
+  async createSurvey(
+    createDto: CreateSurveyDto,
+    user: Users,
+  ): Promise<Surveys> {
     try {
       const { title, description } = createDto;
 
@@ -99,8 +104,12 @@ export class SurveysService {
           '중복된 내용의 설문지가 존재합니다. 다른 내용으로 작성해주세요.',
         );
       }
-      const create = this.surveysRepository.create(createDto);
-      return await this.surveysRepository.save(create);
+      return await this.surveysRepository.save({
+        ...createDto,
+        user,
+      });
+      // const create = this.surveysRepository.create(createDto);
+      // return await this.surveysRepository.save(create);
     } catch (error) {
       this.logger.error(
         `해당 설문지 생성 중 에러가 발생했습니다: ${error.message}`,
@@ -113,9 +122,10 @@ export class SurveysService {
   async updateSurvey(
     surveyId: number,
     updateDto: UpdateSurveyDto,
+    user: Users,
   ): Promise<Surveys> {
     try {
-      await this.surveysRepository.findOneOrFail({
+      const survey = await this.surveysRepository.findOne({
         where: { id: surveyId },
       });
 
@@ -137,6 +147,13 @@ export class SurveysService {
           '중복된 내용의 설문지가 존재합니다. 다른 내용으로 수정해주세요.',
         );
       }
+      if (survey.userId !== user.id) {
+        throw new UnauthorizedException(
+          null,
+          '설문지를 수정할 수 있는 권한이 없습니다.',
+        );
+      }
+
       await this.surveysRepository.update(
         { id: surveyId },
         { title, description },
@@ -151,11 +168,17 @@ export class SurveysService {
   }
 
   // 설문지 삭제 (deleteSurvey)
-  async deleteSurvey(surveyId: number): Promise<EntityWithId> {
+  async deleteSurvey(surveyId: number, user: Users): Promise<EntityWithId> {
     try {
       const survey = await this.surveysRepository.findOneOrFail({
         where: { id: surveyId },
       });
+      if (survey.userId !== user.id) {
+        throw new UnauthorizedException(
+          null,
+          '설문지를 삭제할 수 있는 권한이 없습니다.',
+        );
+      }
       await this.surveysRepository.remove(survey);
       return new EntityWithId(surveyId);
     } catch (error) {
